refactor: move widgets from ChatDetailScreen to seperate files
This makes the chat_detail_screen.dart easier to read and only containing the logic
